Fixed issue #11: Invalid Instance definition.
