+++ back-end/models/order_assignments.js
@@ -21,7 +21,7 @@ module.exports = (sequelize, DataTypes) => {
     },
     user_id: {
       type: DataTypes.INTEGER,
-      autoIncrement: true
+      allowNull: false
     },
     order_id: {
       type: DataTypes.INTEGER,
